refactor: use x-use-component-props instead of useProps
