+++ packages/core/client/src/block-provider/BlockSchemaComponentProvider.tsx
@@ -10,12 +10,13 @@ import { DetailsBlockProvider, useDetailsBlockProps } from './DetailsBlockProvid
 import { FilterFormBlockProvider } from './FilterFormBlockProvider';
 import { FormBlockProvider, useFormBlockProps } from './FormBlockProvider';
 import { FormFieldProvider, useFormFieldProps } from './FormFieldProvider';
-import { TableBlockProvider, useTableBlockProps } from './TableBlockProvider';
+import { TableBlockProvider } from './TableBlockProvider';
+import { useTableBlockProps } from '../modules/blocks/data-blocks/table/hooks/useTableBlockProps';
 import { TableFieldProvider, useTableFieldProps } from './TableFieldProvider';
 import { TableSelectorProvider, useTableSelectorProps } from './TableSelectorProvider';
 import * as bp from './hooks';
 import { BlockSchemaToolbar } from '../modules/blocks/BlockSchemaToolbar';
-import { useTableBlockDecoratorProps } from '../modules/blocks/data-blocks/table/useTableBlockDecoratorProps';
+import { useTableBlockDecoratorProps } from '../modules/blocks/data-blocks/table/hooks/useTableBlockDecoratorProps';
 
 // TODO: delete this, replaced by `BlockSchemaComponentPlugin`
 export const BlockSchemaComponentProvider: React.FC = (props) => {
