+++ packages/core/client/src/block-provider/TableBlockProvider.tsx
@@ -1,13 +1,11 @@
-import { ArrayField, createForm } from '@formily/core';
+import { createForm } from '@formily/core';
 import { FormContext, useField, useFieldSchema } from '@formily/react';
-import React, { createContext, useContext, useEffect, useMemo, useState } from 'react';
+import React, { createContext, useContext, useMemo, useState } from 'react';
 import { useCollectionManager_deprecated } from '../collection-manager';
-import { useFilterBlock } from '../filter-provider/FilterProvider';
-import { mergeFilter } from '../filter-provider/utils';
-import { FixedBlockWrapper, SchemaComponentOptions, removeNullCondition } from '../schema-component';
+import { FixedBlockWrapper, SchemaComponentOptions } from '../schema-component';
 import { BlockProvider, RenderChildrenWithAssociationFilter, useBlockRequestContext } from './BlockProvider';
-import { findFilterTargets, useParsedFilter } from './hooks';
-import { useTableBlockParams } from '../modules/blocks/data-blocks/table/useTableBlockDecoratorProps';
+import { useParsedFilter } from './hooks';
+import { useTableBlockParams } from '../modules/blocks/data-blocks/table/hooks/useTableBlockDecoratorProps';
 import { withDynamicSchemaProps } from '../application/hoc/withDynamicSchemaProps';
 
 export const TableBlockContext = createContext<any>({});
@@ -138,117 +136,4 @@ export const TableBlockProvider = withDynamicSchemaProps((props) => {
 
 export const useTableBlockContext = () => {
   return useContext(TableBlockContext);
-};
-
-export const useTableBlockProps = () => {
-  const field = useField<ArrayField>();
-  const fieldSchema = useFieldSchema();
-  const ctx = useTableBlockContext();
-  const globalSort = fieldSchema.parent?.['x-decorator-props']?.['params']?.['sort'];
-  const { getDataBlocks } = useFilterBlock();
-
-  useEffect(() => {
-    if (!ctx?.service?.loading) {
-      field.value = [];
-      field.value = ctx?.service?.data?.data;
-      field?.setInitialValue(ctx?.service?.data?.data);
-      field.data = field.data || {};
-      field.data.selectedRowKeys = ctx?.field?.data?.selectedRowKeys;
-      field.componentProps.pagination = field.componentProps.pagination || {};
-      field.componentProps.pagination.pageSize = ctx?.service?.data?.meta?.pageSize;
-      field.componentProps.pagination.total = ctx?.service?.data?.meta?.count;
-      field.componentProps.pagination.current = ctx?.service?.data?.meta?.page;
-    }
-  }, [ctx?.service?.data, ctx?.service?.loading]); // 这里如果依赖了 ctx?.field?.data?.selectedRowKeys 的话，会导致这个问题：
-  return {
-    childrenColumnName: ctx.childrenColumnName,
-    loading: ctx?.service?.loading,
-    showIndex: ctx.showIndex,
-    dragSort: ctx.dragSort && ctx.dragSortBy,
-    rowKey: ctx.rowKey || 'id',
-    pagination:
-      ctx?.params?.paginate !== false
-        ? {
-            defaultCurrent: ctx?.params?.page || 1,
-            defaultPageSize: ctx?.params?.pageSize,
-          }
-        : false,
-    onRowSelectionChange(selectedRowKeys) {
-      ctx.field.data = ctx?.field?.data || {};
-      ctx.field.data.selectedRowKeys = selectedRowKeys;
-      ctx?.field?.onRowSelect?.(selectedRowKeys);
-    },
-    async onRowDragEnd({ from, to }) {
-      await ctx.resource.move({
-        sourceId: from[ctx.rowKey || 'id'],
-        targetId: to[ctx.rowKey || 'id'],
-        sortField: ctx.dragSort && ctx.dragSortBy,
-      });
-      ctx.service.refresh();
-    },
-    onChange({ current, pageSize }, filters, sorter) {
-      const sort = sorter.order ? (sorter.order === `ascend` ? [sorter.field] : [`-${sorter.field}`]) : globalSort;
-      ctx.service.run({ ...ctx.service.params?.[0], page: current, pageSize, sort });
-    },
-    onClickRow(record, setSelectedRow, selectedRow) {
-      const { targets, uid } = findFilterTargets(fieldSchema);
-      const dataBlocks = getDataBlocks();
-
-      // 如果是之前创建的区块是没有 x-filter-targets 属性的，所以这里需要判断一下避免报错
-      if (!targets || !targets.some((target) => dataBlocks.some((dataBlock) => dataBlock.uid === target.uid))) {
-        // 当用户已经点击过某一行，如果此时再把相连接的区块给删除的话，行的高亮状态就会一直保留。
-        // 这里暂时没有什么比较好的方法，只是在用户再次点击的时候，把高亮状态给清除掉。
-        setSelectedRow((prev) => (prev.length ? [] : prev));
-        return;
-      }
-
-      const value = [record[ctx.rowKey]];
-
-      dataBlocks.forEach((block) => {
-        const target = targets.find((target) => target.uid === block.uid);
-        if (!target) return;
-
-        const param = block.service.params?.[0] || {};
-        // 保留原有的 filter
-        const storedFilter = block.service.params?.[1]?.filters || {};
-
-        if (selectedRow.includes(record[ctx.rowKey])) {
-          if (block.dataLoadingMode === 'manual') {
-            return block.clearData();
-          }
-          delete storedFilter[uid];
-        } else {
-          storedFilter[uid] = {
-            $and: [
-              {
-                [target.field || ctx.rowKey]: {
-                  [target.field ? '$in' : '$eq']: value,
-                },
-              },
-            ],
-          };
-        }
-
-        const mergedFilter = mergeFilter([
-          ...Object.values(storedFilter).map((filter) => removeNullCondition(filter)),
-          block.defaultFilter,
-        ]);
-
-        return block.doFilter(
-          {
-            ...param,
-            page: 1,
-            filter: mergedFilter,
-          },
-          { filters: storedFilter },
-        );
-      });
-
-      // 更新表格的选中状态
-      setSelectedRow((prev) => (prev?.includes(record[ctx.rowKey]) ? [] : [...value]));
-    },
-    onExpand(expanded, record) {
-      ctx?.field.onExpandClick?.(expanded, record);
-    },
-  };
 };
