+++ packages/core/client/src/modules/blocks/data-blocks/table/__tests__/createTableBLockSchema.test.ts
@@ -1,51 +1,85 @@
 import { createTableBlockUISchema } from '../createTableBlockUISchema';
 
+vi.mock('@formily/shared', () => {
+  return {
+    uid: () => 'mocked-uid',
+  };
+});
+
 describe('createTableBLockSchemaV2', () => {
   it('should create a default table block schema with minimum options', () => {
     const options = { dataSource: 'abc', collectionName: 'users', association: 'users.roles', rowKey: 'rowKey' };
     const schema = createTableBlockUISchema(options);
 
-    expect(schema).toEqual({
-      type: 'void',
-      'x-decorator': 'TableBlockProvider',
-      'x-acl-action': 'users:list',
-      'x-use-decorator-props': 'useTableBlockDecoratorProps',
-      'x-decorator-props': {
-        collection: 'users',
-        dataSource: 'abc',
-        association: 'users.roles',
-        action: 'list',
-        params: {
-          pageSize: 20,
+    expect(schema).toMatchInlineSnapshot(`
+      {
+        "properties": {
+          "actions": {
+            "properties": {},
+            "type": "void",
+            "x-component": "ActionBar",
+            "x-component-props": {
+              "style": {
+                "marginBottom": "var(--nb-spacing)",
+              },
+            },
+            "x-initializer": "table:configureActions",
+          },
+          "mocked-uid": {
+            "properties": {
+              "actions": {
+                "properties": {
+                  "mocked-uid": {
+                    "type": "void",
+                    "x-component": "Space",
+                    "x-component-props": {
+                      "split": "|",
+                    },
+                    "x-decorator": "DndContext",
+                  },
+                },
+                "title": "{{ t("Actions") }}",
+                "type": "void",
+                "x-action-column": "actions",
+                "x-component": "TableV2.Column",
+                "x-decorator": "TableV2.Column.ActionBar",
+                "x-designer": "TableV2.ActionColumnDesigner",
+                "x-initializer": "table:configureItemActions",
+              },
+            },
+            "type": "array",
+            "x-component": "TableV2",
+            "x-component-props": {
+              "rowKey": "id",
+              "rowSelection": {
+                "type": "checkbox",
+              },
+            },
+            "x-initializer": "table:configureColumns",
+            "x-use-component-props": "useTableBlockProps",
+          },
         },
-        rowKey: 'rowKey',
-        showIndex: true,
-        dragSort: false,
-      },
-      'x-toolbar': 'BlockSchemaToolbar',
-      'x-settings': 'blockSettings:table',
-      'x-component': 'CardItem',
-      'x-filter-targets': [],
-      properties: expect.any(Object), // since uid() generates a unique value every time, we check for any Object.
-    });
-  });
-
-  it('should handle custom dataSource, rowKey, and collectionName', () => {
-    const options = {
-      collectionName: 'products',
-      dataSource: 'productDataSource',
-      rowKey: 'productId',
-      association: 'productAssociation.name',
-    };
-    const schema = createTableBlockUISchema(options);
-
-    expect(schema['x-decorator-props']).toEqual(
-      expect.objectContaining({
-        collection: 'products',
-        dataSource: 'productDataSource',
-        rowKey: 'productId',
-        association: 'productAssociation.name',
-      }),
-    );
+        "type": "void",
+        "x-acl-action": "users:list",
+        "x-component": "CardItem",
+        "x-decorator": "TableBlockProvider",
+        "x-decorator-props": {
+          "action": "list",
+          "association": "users.roles",
+          "collection": "users",
+          "dataSource": "abc",
+          "dragSort": false,
+          "params": {
+            "pageSize": 20,
+          },
+          "rowKey": "rowKey",
+          "showIndex": true,
+        },
+        "x-filter-targets": [],
+        "x-settings": "blockSettings:table",
+        "x-toolbar": "BlockSchemaToolbar",
+        "x-use-decorator-props": "useTableBlockDecoratorProps",
+      }
+    `);
   });
 });
