+++ packages/core/client/src/modules/blocks/data-blocks/table/createTableBlockUISchema.ts
@@ -50,12 +50,12 @@ export const createTableBlockUISchema = (options: {
         type: 'array',
         'x-initializer': 'table:configureColumns',
         'x-component': 'TableV2',
+        'x-use-component-props': 'useTableBlockProps',
         'x-component-props': {
           rowKey: 'id',
           rowSelection: {
             type: 'checkbox',
           },
-          useProps: '{{ useTableBlockProps }}',
         },
         properties: {
           actions: {
