+++ packages/core/client/src/schema-component/antd/table-v2/Table.tsx
@@ -24,6 +24,7 @@ import {
   useSchemaInitializerRender,
   useTableBlockContext,
   useTableSelectorContext,
+  withDynamicSchemaProps,
 } from '../../../';
 import { useACLFieldWhitelist } from '../../../acl/ACLProvider';
 import { useToken } from '../__builtins__';
@@ -212,384 +213,390 @@ const usePaginationProps = (pagination1, pagination2) => {
 };
 
 export const Table: any = observer(
-  (props: {
-    useProps?: () => any;
-    onChange?: (pagination, filters, sorter, extra) => void;
-    onRowSelectionChange?: (selectedRowKeys: any[], selectedRows: any[]) => void;
-    onRowDragEnd?: (e: { from: any; to: any }) => void;
-    onClickRow?: (record: any, setSelectedRow: (selectedRow: any[]) => void, selectedRow: any[]) => void;
-    pagination?: any;
-    showIndex?: boolean;
-    dragSort?: boolean;
-    rowKey?: string | ((record: any) => string);
-    rowSelection?: any;
-    required?: boolean;
-    onExpand?: (flag: boolean, record: any) => void;
-    isSubTable?: boolean;
-  }) => {
-    const { token } = useToken();
-    const { pagination: pagination1, useProps, onChange, ...others1 } = props;
-    const { pagination: pagination2, onClickRow, ...others2 } = useProps?.() || {};
-    const {
-      dragSort = false,
-      showIndex = true,
-      onRowSelectionChange,
-      onChange: onTableChange,
-      rowSelection,
-      rowKey,
-      required,
-      onExpand,
-      ...others
-    } = { ...others1, ...others2 } as any;
-    const field = useArrayField(others);
-    const columns = useTableColumns(others);
-    const schema = useFieldSchema();
-    const collection = useCollection_deprecated();
-    const isTableSelector = schema?.parent?.['x-decorator'] === 'TableSelectorProvider';
-    const ctx = isTableSelector ? useTableSelectorContext() : useTableBlockContext();
-    const { expandFlag, allIncludesChildren } = ctx;
-    const onRowDragEnd = useMemoizedFn(others.onRowDragEnd || (() => {}));
-    const paginationProps = usePaginationProps(pagination1, pagination2);
-    const [expandedKeys, setExpandesKeys] = useState([]);
-    const [selectedRowKeys, setSelectedRowKeys] = useState<any[]>(field?.data?.selectedRowKeys || []);
-    const [selectedRow, setSelectedRow] = useState([]);
-    const dataSource = field?.value?.slice?.()?.filter?.(Boolean) || [];
-    const isRowSelect = rowSelection?.type !== 'none';
-    const defaultRowKeyMap = useRef(new Map());
-    let onRow = null,
-      highlightRow = '';
+  withDynamicSchemaProps(
+    (props: {
+      useProps?: () => any;
+      onChange?: (pagination, filters, sorter, extra) => void;
+      onRowSelectionChange?: (selectedRowKeys: any[], selectedRows: any[]) => void;
+      onRowDragEnd?: (e: { from: any; to: any }) => void;
+      onClickRow?: (record: any, setSelectedRow: (selectedRow: any[]) => void, selectedRow: any[]) => void;
+      pagination?: any;
+      showIndex?: boolean;
+      dragSort?: boolean;
+      rowKey?: string | ((record: any) => string);
+      rowSelection?: any;
+      required?: boolean;
+      onExpand?: (flag: boolean, record: any) => void;
+      isSubTable?: boolean;
+    }) => {
+      const { token } = useToken();
+      const { pagination: pagination1, useProps, onChange, ...others1 } = props;
 
-    if (onClickRow) {
-      onRow = (record) => {
-        return {
-          onClick: (e) => {
-            if (isPortalInBody(e.target)) {
-              return;
-            }
-            onClickRow(record, setSelectedRow, selectedRow);
-          },
+      // 新版 UISchema（1.0 之后）中已经废弃了 useProps，这里之所以继续保留是为了兼容旧版的 UISchema
+      const { pagination: pagination2, ...others2 } = useProps?.() || {};
+
+      const {
+        dragSort = false,
+        showIndex = true,
+        onRowSelectionChange,
+        onChange: onTableChange,
+        rowSelection,
+        rowKey,
+        required,
+        onExpand,
+        onClickRow,
+        ...others
+      } = { ...others1, ...others2 } as any;
+      const field = useArrayField(others);
+      const columns = useTableColumns(others);
+      const schema = useFieldSchema();
+      const collection = useCollection_deprecated();
+      const isTableSelector = schema?.parent?.['x-decorator'] === 'TableSelectorProvider';
+      const ctx = isTableSelector ? useTableSelectorContext() : useTableBlockContext();
+      const { expandFlag, allIncludesChildren } = ctx;
+      const onRowDragEnd = useMemoizedFn(others.onRowDragEnd || (() => {}));
+      const paginationProps = usePaginationProps(pagination1, pagination2);
+      const [expandedKeys, setExpandesKeys] = useState([]);
+      const [selectedRowKeys, setSelectedRowKeys] = useState<any[]>(field?.data?.selectedRowKeys || []);
+      const [selectedRow, setSelectedRow] = useState([]);
+      const dataSource = field?.value?.slice?.()?.filter?.(Boolean) || [];
+      const isRowSelect = rowSelection?.type !== 'none';
+      const defaultRowKeyMap = useRef(new Map());
+      let onRow = null,
+        highlightRow = '';
+
+      if (onClickRow) {
+        onRow = (record) => {
+          return {
+            onClick: (e) => {
+              if (isPortalInBody(e.target)) {
+                return;
+              }
+              onClickRow(record, setSelectedRow, selectedRow);
+            },
+          };
         };
-      };
-      highlightRow = css`
-        & > td {
-          background-color: ${token.controlItemBgActiveHover} !important;
-        }
-        &:hover > td {
-          background-color: ${token.controlItemBgActiveHover} !important;
-        }
-      `;
-    }
-
-    useEffect(() => {
-      if (expandFlag) {
-        setExpandesKeys(allIncludesChildren);
-      } else {
-        setExpandesKeys([]);
+        highlightRow = css`
+          & > td {
+            background-color: ${token.controlItemBgActiveHover} !important;
+          }
+          &:hover > td {
+            background-color: ${token.controlItemBgActiveHover} !important;
+          }
+        `;
       }
-    }, [expandFlag, allIncludesChildren]);
 
-    const components = useMemo(() => {
-      return {
-        header: {
-          wrapper: (props) => {
-            return (
-              <DndContext>
-                <thead {...props} />
-              </DndContext>
-            );
+      useEffect(() => {
+        if (expandFlag) {
+          setExpandesKeys(allIncludesChildren);
+        } else {
+          setExpandesKeys([]);
+        }
+      }, [expandFlag, allIncludesChildren]);
+
+      const components = useMemo(() => {
+        return {
+          header: {
+            wrapper: (props) => {
+              return (
+                <DndContext>
+                  <thead {...props} />
+                </DndContext>
+              );
+            },
+            cell: (props) => {
+              return (
+                <th
+                  {...props}
+                  className={cls(
+                    props.className,
+                    css`
+                      max-width: 300px;
+                      white-space: nowrap;
+                      &:hover .general-schema-designer {
+                        display: block;
+                      }
+                    `,
+                  )}
+                />
+              );
+            },
           },
-          cell: (props) => {
-            return (
-              <th
+          body: {
+            wrapper: (props) => {
+              return (
+                <DndContext
+                  onDragEnd={(e) => {
+                    if (!e.active || !e.over) {
+                      console.warn('move cancel');
+                      return;
+                    }
+                    const fromIndex = e.active?.data.current?.sortable?.index;
+                    const toIndex = e.over?.data.current?.sortable?.index;
+                    const from = field.value[fromIndex] || e.active;
+                    const to = field.value[toIndex] || e.over;
+                    void field.move(fromIndex, toIndex);
+                    onRowDragEnd({ from, to });
+                  }}
+                >
+                  <tbody {...props} />
+                </DndContext>
+              );
+            },
+            row: (props) => {
+              return <SortableRow {...props}></SortableRow>;
+            },
+            cell: (props) => (
+              <td
                 {...props}
-                className={cls(
+                className={classNames(
                   props.className,
                   css`
                     max-width: 300px;
                     white-space: nowrap;
-                    &:hover .general-schema-designer {
-                      display: block;
+                    .nb-read-pretty-input-number {
+                      text-align: right;
+                    }
+                    .ant-color-picker-trigger {
+                      position: absolute;
+                      top: 50%;
+                      transform: translateY(-50%);
                     }
                   `,
                 )}
               />
-            );
+            ),
           },
-        },
-        body: {
-          wrapper: (props) => {
-            return (
-              <DndContext
-                onDragEnd={(e) => {
-                  if (!e.active || !e.over) {
-                    console.warn('move cancel');
-                    return;
-                  }
-                  const fromIndex = e.active?.data.current?.sortable?.index;
-                  const toIndex = e.over?.data.current?.sortable?.index;
-                  const from = field.value[fromIndex] || e.active;
-                  const to = field.value[toIndex] || e.over;
-                  void field.move(fromIndex, toIndex);
-                  onRowDragEnd({ from, to });
-                }}
-              >
-                <tbody {...props} />
-              </DndContext>
-            );
-          },
-          row: (props) => {
-            return <SortableRow {...props}></SortableRow>;
-          },
-          cell: (props) => (
-            <td
-              {...props}
-              className={classNames(
-                props.className,
-                css`
-                  max-width: 300px;
-                  white-space: nowrap;
-                  .nb-read-pretty-input-number {
-                    text-align: right;
-                  }
-                  .ant-color-picker-trigger {
-                    position: absolute;
-                    top: 50%;
-                    transform: translateY(-50%);
-                  }
-                `,
-              )}
-            />
-          ),
-        },
+        };
+      }, [field, onRowDragEnd, dragSort]);
+
+      /**
+       * 为没有设置 key 属性的表格行生成一个唯一的 key
+       * 1. rowKey 的默认值是 “key”，所以先判断有没有 record.key；
+       * 2. 如果没有就生成一个唯一的 key，并以 record 的值作为索引；
+       * 3. 这样下次就能取到对应的 key 的值；
+       *
+       * 这里有效的前提是：数组中对应的 record 的引用不会发生改变。
+       *
+       * @param record
+       * @returns
+       */
+      const defaultRowKey = (record: any) => {
+        if (record.key) {
+          return record.key;
+        }
+
+        if (defaultRowKeyMap.current.has(record)) {
+          return defaultRowKeyMap.current.get(record);
+        }
+
+        const key = uid();
+        defaultRowKeyMap.current.set(record, key);
+        return key;
       };
-    }, [field, onRowDragEnd, dragSort]);
 
-    /**
-     * 为没有设置 key 属性的表格行生成一个唯一的 key
-     * 1. rowKey 的默认值是 “key”，所以先判断有没有 record.key；
-     * 2. 如果没有就生成一个唯一的 key，并以 record 的值作为索引；
-     * 3. 这样下次就能取到对应的 key 的值；
-     *
-     * 这里有效的前提是：数组中对应的 record 的引用不会发生改变。
-     *
-     * @param record
-     * @returns
-     */
-    const defaultRowKey = (record: any) => {
-      if (record.key) {
-        return record.key;
-      }
+      const getRowKey = (record: any) => {
+        if (typeof rowKey === 'string') {
+          return record[rowKey]?.toString();
+        } else {
+          return (rowKey ?? defaultRowKey)(record)?.toString();
+        }
+      };
 
-      if (defaultRowKeyMap.current.has(record)) {
-        return defaultRowKeyMap.current.get(record);
-      }
-
-      const key = uid();
-      defaultRowKeyMap.current.set(record, key);
-      return key;
-    };
-
-    const getRowKey = (record: any) => {
-      if (typeof rowKey === 'string') {
-        return record[rowKey]?.toString();
-      } else {
-        return (rowKey ?? defaultRowKey)(record)?.toString();
-      }
-    };
-
-    const restProps = {
-      rowSelection: rowSelection
-        ? {
-            type: 'checkbox',
-            selectedRowKeys: selectedRowKeys,
-            onChange(selectedRowKeys: any[], selectedRows: any[]) {
-              field.data = field.data || {};
-              field.data.selectedRowKeys = selectedRowKeys;
-              setSelectedRowKeys(selectedRowKeys);
-              onRowSelectionChange?.(selectedRowKeys, selectedRows);
-            },
-            getCheckboxProps(record) {
-              return {
-                'aria-label': `checkbox`,
-              };
-            },
-            renderCell: (checked, record, index, originNode) => {
-              if (!dragSort && !showIndex) {
-                return originNode;
-              }
-              const current = props?.pagination?.current;
-              const pageSize = props?.pagination?.pageSize || 20;
-              if (current) {
-                index = index + (current - 1) * pageSize + 1;
-              } else {
-                index = index + 1;
-              }
-              if (record.__index) {
-                index = extractIndex(record.__index);
-              }
-              return (
-                <div
-                  role="button"
-                  aria-label={`table-index-${index}`}
-                  className={classNames(
-                    checked ? 'checked' : null,
-                    css`
-                      position: relative;
-                      display: flex;
-                      float: left;
-                      align-items: center;
-                      justify-content: space-evenly;
-                      padding-right: 8px;
-                      .nb-table-index {
-                        opacity: 0;
-                      }
-                      &:not(.checked) {
-                        .nb-table-index {
-                          opacity: 1;
-                        }
-                      }
-                    `,
-                    {
-                      [css`
-                        &:hover {
-                          .nb-table-index {
-                            opacity: 0;
-                          }
-                          .nb-origin-node {
-                            display: block;
-                          }
-                        }
-                      `]: isRowSelect,
-                    },
-                  )}
-                >
+      const restProps = {
+        rowSelection: rowSelection
+          ? {
+              type: 'checkbox',
+              selectedRowKeys: selectedRowKeys,
+              onChange(selectedRowKeys: any[], selectedRows: any[]) {
+                field.data = field.data || {};
+                field.data.selectedRowKeys = selectedRowKeys;
+                setSelectedRowKeys(selectedRowKeys);
+                onRowSelectionChange?.(selectedRowKeys, selectedRows);
+              },
+              getCheckboxProps(record) {
+                return {
+                  'aria-label': `checkbox`,
+                };
+              },
+              renderCell: (checked, record, index, originNode) => {
+                if (!dragSort && !showIndex) {
+                  return originNode;
+                }
+                const current = props?.pagination?.current;
+                const pageSize = props?.pagination?.pageSize || 20;
+                if (current) {
+                  index = index + (current - 1) * pageSize + 1;
+                } else {
+                  index = index + 1;
+                }
+                if (record.__index) {
+                  index = extractIndex(record.__index);
+                }
+                return (
                   <div
+                    role="button"
+                    aria-label={`table-index-${index}`}
                     className={classNames(
                       checked ? 'checked' : null,
                       css`
                         position: relative;
                         display: flex;
+                        float: left;
                         align-items: center;
                         justify-content: space-evenly;
+                        padding-right: 8px;
+                        .nb-table-index {
+                          opacity: 0;
+                        }
+                        &:not(.checked) {
+                          .nb-table-index {
+                            opacity: 1;
+                          }
+                        }
                       `,
+                      {
+                        [css`
+                          &:hover {
+                            .nb-table-index {
+                              opacity: 0;
+                            }
+                            .nb-origin-node {
+                              display: block;
+                            }
+                          }
+                        `]: isRowSelect,
+                      },
                     )}
                   >
-                    {dragSort && <SortHandle id={getRowKey(record)} />}
-                    {showIndex && <TableIndex index={index} />}
-                  </div>
-                  {isRowSelect && (
                     <div
                       className={classNames(
-                        'nb-origin-node',
                         checked ? 'checked' : null,
                         css`
-                          position: absolute;
-                          right: 50%;
-                          transform: translateX(50%);
-                          &:not(.checked) {
-                            display: none;
-                          }
+                          position: relative;
+                          display: flex;
+                          align-items: center;
+                          justify-content: space-evenly;
                         `,
                       )}
                     >
-                      {originNode}
+                      {dragSort && <SortHandle id={getRowKey(record)} />}
+                      {showIndex && <TableIndex index={index} />}
                     </div>
-                  )}
-                </div>
-              );
-            },
-            ...rowSelection,
-          }
-        : undefined,
-    };
-    const SortableWrapper = useCallback<React.FC>(
-      ({ children }) => {
-        return dragSort
-          ? React.createElement<Omit<SortableContextProps, 'children'>>(
-              SortableContext,
-              {
-                items: field.value?.map?.(getRowKey) || [],
+                    {isRowSelect && (
+                      <div
+                        className={classNames(
+                          'nb-origin-node',
+                          checked ? 'checked' : null,
+                          css`
+                            position: absolute;
+                            right: 50%;
+                            transform: translateX(50%);
+                            &:not(.checked) {
+                              display: none;
+                            }
+                          `,
+                        )}
+                      >
+                        {originNode}
+                      </div>
+                    )}
+                  </div>
+                );
               },
-              children,
-            )
-          : React.createElement(React.Fragment, {}, children);
-      },
-      [field, dragSort],
-    );
-    const fieldSchema = useFieldSchema();
-    const fixedBlock = fieldSchema?.parent?.['x-decorator-props']?.fixedBlock;
+              ...rowSelection,
+            }
+          : undefined,
+      };
+      const SortableWrapper = useCallback<React.FC>(
+        ({ children }) => {
+          return dragSort
+            ? React.createElement<Omit<SortableContextProps, 'children'>>(
+                SortableContext,
+                {
+                  items: field.value?.map?.(getRowKey) || [],
+                },
+                children,
+              )
+            : React.createElement(React.Fragment, {}, children);
+        },
+        [field, dragSort],
+      );
+      const fieldSchema = useFieldSchema();
+      const fixedBlock = fieldSchema?.parent?.['x-decorator-props']?.fixedBlock;
 
-    const { height: tableHeight, tableSizeRefCallback } = useTableSize();
-    const scroll = useMemo(() => {
-      return fixedBlock
-        ? {
-            x: 'max-content',
-            y: tableHeight,
-          }
-        : {
-            x: 'max-content',
-          };
-    }, [fixedBlock, tableHeight]);
-    return (
-      <div
-        className={css`
-          height: 100%;
-          overflow: hidden;
-          .ant-table-wrapper {
+      const { height: tableHeight, tableSizeRefCallback } = useTableSize();
+      const scroll = useMemo(() => {
+        return fixedBlock
+          ? {
+              x: 'max-content',
+              y: tableHeight,
+            }
+          : {
+              x: 'max-content',
+            };
+      }, [fixedBlock, tableHeight]);
+      return (
+        <div
+          className={css`
             height: 100%;
-            .ant-spin-nested-loading {
+            overflow: hidden;
+            .ant-table-wrapper {
               height: 100%;
-              .ant-spin-container {
+              .ant-spin-nested-loading {
                 height: 100%;
-                display: flex;
-                flex-direction: column;
+                .ant-spin-container {
+                  height: 100%;
+                  display: flex;
+                  flex-direction: column;
+                }
               }
             }
-          }
-          .ant-table {
-            overflow-x: auto;
-            overflow-y: hidden;
-          }
-        `}
-      >
-        <SortableWrapper>
-          <AntdTable
-            ref={tableSizeRefCallback}
-            rowKey={rowKey ?? defaultRowKey}
-            dataSource={dataSource}
-            tableLayout="auto"
-            {...others}
-            {...restProps}
-            pagination={paginationProps}
-            components={components}
-            onChange={(pagination, filters, sorter, extra) => {
-              onTableChange?.(pagination, filters, sorter, extra);
-            }}
-            onRow={onRow}
-            rowClassName={(record) => (selectedRow.includes(record[rowKey]) ? highlightRow : '')}
-            scroll={scroll}
-            columns={columns}
-            expandable={{
-              onExpand: (flag, record) => {
-                const newKeys = flag
-                  ? [...expandedKeys, record[collection.getPrimaryKey()]]
-                  : expandedKeys.filter((i) => record[collection.getPrimaryKey()] !== i);
-                setExpandesKeys(newKeys);
-                onExpand?.(flag, record);
-              },
-              expandedRowKeys: expandedKeys,
-            }}
-          />
-        </SortableWrapper>
-        {field.errors.length > 0 && (
-          <div className="ant-formily-item-error-help ant-formily-item-help ant-formily-item-help-enter ant-formily-item-help-enter-active">
-            {field.errors.map((error) => {
-              return error.messages.map((message) => <div key={message}>{message}</div>);
-            })}
-          </div>
-        )}
-      </div>
-    );
-  },
+            .ant-table {
+              overflow-x: auto;
+              overflow-y: hidden;
+            }
+          `}
+        >
+          <SortableWrapper>
+            <AntdTable
+              ref={tableSizeRefCallback}
+              rowKey={rowKey ?? defaultRowKey}
+              dataSource={dataSource}
+              tableLayout="auto"
+              {...others}
+              {...restProps}
+              pagination={paginationProps}
+              components={components}
+              onChange={(pagination, filters, sorter, extra) => {
+                onTableChange?.(pagination, filters, sorter, extra);
+              }}
+              onRow={onRow}
+              rowClassName={(record) => (selectedRow.includes(record[rowKey]) ? highlightRow : '')}
+              scroll={scroll}
+              columns={columns}
+              expandable={{
+                onExpand: (flag, record) => {
+                  const newKeys = flag
+                    ? [...expandedKeys, record[collection.getPrimaryKey()]]
+                    : expandedKeys.filter((i) => record[collection.getPrimaryKey()] !== i);
+                  setExpandesKeys(newKeys);
+                  onExpand?.(flag, record);
+                },
+                expandedRowKeys: expandedKeys,
+              }}
+            />
+          </SortableWrapper>
+          {field.errors.length > 0 && (
+            <div className="ant-formily-item-error-help ant-formily-item-help ant-formily-item-help-enter ant-formily-item-help-enter-active">
+              {field.errors.map((error) => {
+                return error.messages.map((message) => <div key={message}>{message}</div>);
+              })}
+            </div>
+          )}
+        </div>
+      );
+    },
+  ),
   { displayName: 'Table' },
 );
